Revert "Separate interfaces for separated files"
This reverts commit a44f888978b9ab33296946a7288e4c8c27c4427a.
