--- src/redux/projects/types/index.ts
@@ -1,4 +1,0 @@
-export * from "./add.type";
-export * from "./remove.type";
-export * from "./edit.type";
-export * from "./global.type";
